requirements.txt
aioredis==1.3.1
asgiref==3.2.7
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==3.1.0
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
Django==3.0.3
django-bootstrap==0.2.4
django-bootstrap4==2.2.0
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.11.1
djongo==1.3.3
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
msgpack==1.0.0
optional-django==0.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pymongo==3.11.0
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
Twisted==20.3.0
txaio==20.4.1
wincertstore==0.2
zope.interface==5.1.2
